+++ src/backtracking.hpp
@@ -6,15 +6,15 @@
 
 class Backtracking : public TSPStrategy {
 public:
-  virtual uint64_t solve(Graph *graph,
-                         std::optional<uint64_t> entry_value = std::nullopt);
+  virtual uint16_t solve(Graph *graph,
+                         std::optional<uint16_t> entry_value = std::nullopt);
 
 private:
-  void helper(uint64_t pos, uint64_t count, uint64_t dist);
+  void helper(uint8_t pos, uint16_t count, uint16_t dist);
 
   Graph *graph;
   std::vector<bool> visited;
-  uint64_t *ans;
+  uint16_t *ans;
 };
 
 #endif // TSP_BACKTRACKING
