+++ src/bruteforce.cc
@@ -3,16 +3,16 @@
 #include <limits>
 #include <vector>
 
-uint64_t Bruteforce::solve(Graph *graph, std::optional<uint64_t> entry_value) {
-  std::vector<uint64_t> vertices(graph->size() - 1);
+uint16_t Bruteforce::solve(Graph *graph, std::optional<uint16_t> entry_value) {
+  std::vector<uint8_t> vertices(graph->size() - 1);
   for (int i = 0; i < vertices.size(); i++)
     vertices[i] = i;
 
-  uint64_t result = (entry_value.has_value())
+  uint16_t result = (entry_value.has_value())
                         ? entry_value.value()
-                        : std::numeric_limits<uint64_t>::max();
+                        : std::numeric_limits<uint16_t>::max();
   do {
-    uint64_t total = graph->distance(0, vertices[vertices.size() - 1] + 1) +
+    uint16_t total = graph->distance(0, vertices[vertices.size() - 1] + 1) +
                      graph->distance(0, vertices[0] + 1);
     int i = 0;
     for (; i < vertices.size() - 1 && total < result; i++)
