+++ src/bruteforce_cuda.hpp
@@ -5,8 +5,8 @@
 
 class BruteforceCUDA : public TSPStrategy {
 public:
-  virtual uint64_t solve(Graph *graph,
-                         std::optional<uint64_t> entry_value = std::nullopt);
+  virtual uint16_t solve(Graph *graph,
+                         std::optional<uint16_t> entry_value = std::nullopt);
 };
 
 #endif // TSP_BRUTEFORCE_CUDA
