+++ src/graph.hpp
@@ -9,18 +9,18 @@ public:
   Graph(nlohmann::json data);
 
   std::size_t size();
-  uint64_t distance(uint64_t from, uint64_t to);
+  uint16_t distance(uint8_t from, uint8_t to);
   std::string print();
 
 private:
-  std::vector<std::vector<uint64_t>> distances;
+  std::vector<std::vector<uint16_t>> distances;
 };
 
 class TSPStrategy {
 public:
   virtual ~TSPStrategy() {}
-  virtual uint64_t
-  solve(Graph *graph, std::optional<uint64_t> entry_value = std::nullopt) = 0;
+  virtual uint16_t
+  solve(Graph *graph, std::optional<uint16_t> entry_value = std::nullopt) = 0;
 };
 
 #endif // TSP_GRAPH
